RR-173: Fixed other code review issues
