+++ src/client/actions.js
@@ -35,9 +35,6 @@ export const EVENTS__LOADED = 'EVENTS__LOADED'
 export const EVENTS__METADATA_LOADED = 'EVENTS__METADATA_LOADED'
 export const EVENTS__SELECTED_ORGANISER = 'EVENTS__SELECTED_ORGANISER'
 export const EVENTS__DETAILS_LOADED = 'EVENTS__DETAILS_LOADED'
-export const EVENTS__FORM_METADATA_LOADED = 'EVENTS__FORM_METADATA_LOADED'
-
-export const ADD_EVENT_FORM__SUBMIT = 'ADD_EVENT_FORM__SUBMIT'
 
 export const REFERRAL_DETAILS = 'REFERRAL_DETAILS'
