--- src/client/components/Resource/Event.js
@@ -1,3 +1,0 @@
-import { createEntityResource } from '.'
-
-export default createEntityResource('Event', (id) => `v4/event/${id}`)
